media: bcm2835-unicam: Do not write new buffer address on Frame End
There is a possible race condition when programming the buffer
address on frame end.  Unicam may have already seen the FS event on the
next frame, and will not latch the buffer until the *next* FS event,
but the software assumes it will.

Fix so that we can program the buffer up until the last line interrupt.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>