media: bcm2835-unicam: Add a mechanism to pass out generic metadata.
This will be used in cases where the sensor does not support
embedded data output.  We still advertise node[1] as an embedded
data node, and get buffers queued as normal.  On Frame End, we write
the exposure and gain values to the queued buffers directly from our
delayed ctrl list.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>